pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lana-sim-scenarios"
version = "0.1.0"
description = "Convert YAML scenarios to lana-bank sim-bootstrap Rust code"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "black>=23.0",
]

[project.scripts]
lana-sim = "lana_sim_scenarios.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
